Tighten the types the upgraded analysers now check
Console validators receive whatever the user typed, so they take mixed and
prove the type themselves; Length takes named arguments instead of an options
array; getUserIdentifier states the non-empty string Symfony requires;
convertCodeToName never returns null, as its catch falls back to the code.
The rest are annotations that no longer matched the values behind them.

+++ src/CongregationManager/Bundle/TerritoryManager/src/Renderer/WordS13Renderer.php
@@ -182,7 +182,7 @@ final readonly class WordS13Renderer implements S13RendererInterface
                 ];
             }
             $assignmentsTableAssignedToColumn = $assignmentsTable->addCell(
-                Converter::cmToTwip(self::DATE_ASSIGNED_COLUMN_CM + self::DATE_COMPLETED_COLUMN_CM),
+                Converter::cmToTwip((float) self::DATE_ASSIGNED_COLUMN_CM + self::DATE_COMPLETED_COLUMN_CM),
                 array_merge($lastColumnParagraphStyle, [
                     'gridSpan' => 2,
                     'valign' => TextAlignment::CENTER,
@@ -310,7 +310,7 @@ final readonly class WordS13Renderer implements S13RendererInterface
                 ];
             }
             $assignmentsTableAssignedToColumn = $assignmentsTable->addCell(
-                Converter::cmToTwip(self::DATE_ASSIGNED_COLUMN_CM + self::DATE_COMPLETED_COLUMN_CM),
+                Converter::cmToTwip((float) self::DATE_ASSIGNED_COLUMN_CM + self::DATE_COMPLETED_COLUMN_CM),
                 array_merge($lastColumnParagraphStyle, [
                     'gridSpan' => 2,
                     'valign' => TextAlignment::CENTER,
